Refactor code into modular components
- Extract models (MemoryUnit, TableRow, TreeNode) to models.py
- Move table formatting logic to table_formatter.py
- Create tree_builder.py for hierarchical key structure building
- Update CLI to use new modular structure
- Add comprehensive tests for each module
